+++ src/commands/hardis/org/configure/files.ts
@@ -63,6 +63,7 @@ See article below
       soqlQuery: "SELECT Id,Name FROM Opportunity",
       fileTypes: "all",
       outputFolderNameField: "Name",
+      outputFileNameFormat: "title",
       overwriteParentRecords: true,
       overwriteFiles: false,
     };