+++ src/common/utils/wrapUtils.ts
@@ -53,6 +53,11 @@ export async function wrapSfdxCoreCommand(commandBase: string, argv: string[], c
       uxLog(this, c.red("\n" + errLog));
     }
     deployRes = errLog;
+    if (e.code) {
+      process.exitCode = e.code;
+    } else {
+      process.exitCode = 1;
+    }
   }
   return { outputstring: deployRes };
 }